Implement breadcrumbs
There is a tiny hack in the breadcrumb impl, which is that we are checking if a page should be included by checking if it contains rather than if it prefixes.
Would be possible to fix, but requires additional logic and only matters if we create another page named "about", "docs", "projects", or "posts". For now lets not bother fixing that as that seems unlikely